+++ packages/terra-button-group/src/terra-dev-site/test/button-group/ButtonGroupWrapping.test.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import ButtonGroup from '../../../ButtonGroup';
+
+const ButtonGroupWrapping = () => (
+  <ButtonGroup id="button-group-text">
+    <ButtonGroup.Button text="A-Really-Long Text" key="text1" />
+    <ButtonGroup.Button text="A-Really-Long Text" key="text2" />
+    <ButtonGroup.Button text="A-Really-Long Text" key="text3" />
+    <ButtonGroup.Button text="A-Really-Long Text" key="text4" />
+    <ButtonGroup.Button text="A-Really-Long Text" key="text5" />
+  </ButtonGroup>
+);
+
+export default ButtonGroupWrapping;