+++ packages/terra-show-hide/src/terra-dev-site/doc/ShowHide/ChangeLog.2.doc.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import DocTemplate from 'terra-doc-template';
+import ChangeLog from '../../../../CHANGELOG.md';
+
+const DocPage = () => (
+  <DocTemplate
+    packageName="show-hide"
+    srcPath="https://github.com/cerner/terra-core/tree/master/packages/show-hide"
+    readme={ChangeLog}
+  />
+);
+
+export default DocPage;