+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -4,6 +4,21 @@ ChangeLog
 Unreleased
 ----------
 
+2.34.0 - (October 16, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+2.33.0 - (October 8, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+2.32.0 - (October 2, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.31.0 - (September 25, 2018)
 ------------------
 ### Changed