+++ packages/terra-form-field/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+2.27.0 - (October 16, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+2.26.0 - (October 8, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.25.0 - (September 25, 2018)
 ------------------
 ### Changed