+++ packages/terra-form-select/CHANGELOG.md
@@ -4,6 +4,22 @@ ChangeLog
 Unreleased
 ----------
 
+4.20.0 - (October 16, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+### Added
+* Theme variable for input height
+
+### Fixed
+* Normalized duplicate padding theme variable
+
+4.19.0 - (October 8, 2018)
+------------------
+### Added
+* First class prop support and documentation for onBlur, onFocus, and onClick
+
 4.18.0 - (September 25, 2018)
 ------------------
 ### Changed