+++ packages/terra-hyperlink/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+1.13.0 - (October 16, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+1.12.0 - (October 8, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
 1.11.0 - (September 25, 2018)
 ------------------
 ### Changed