+++ packages/terra-icon/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+2.25.0 - (October 16, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+2.24.0 - (October 8, 2018)
+------------------
+### Changed
+* Updated to use one-cerner-style-icons 1.19.2
+
 2.23.0 - (September 25, 2018)
 ------------------
 ### Changed