+++ packages/terra-markdown/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+2.15.0 - (October 8, 2018)
+------------------
+### Changed
+* Updated links in docs on supported features
+
 2.14.0 - (September 5, 2018)
 ------------------
 ### Changed