+++ packages/terra-menu/CHANGELOG.md
@@ -4,6 +4,21 @@ Changelog
 Unreleased
 ----------
 
+3.23.0 - (October 16, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.22.0 - (October 8, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.21.0 - (October 2, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
 3.20.0 - (September 25, 2018)
 ------------------
 ### Changed