+++ packages/terra-paginator/CHANGELOG.md
@@ -4,6 +4,21 @@ ChangeLog
 Unreleased
 ----------
 
+1.21.0 - (October 16, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+1.20.0 - (October 8, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+1.19.0 - (October 2, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
 1.18.0 - (September 25, 2018)
 ------------------
 ### Changed