+++ packages/terra-responsive-element/CHANGELOG.md
@@ -3,6 +3,19 @@ Changelog
 
 Unreleased
 ----------
+
+3.13.0 - (October 16, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.12.0 - (October 8, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+3.11.0 - (October 2, 2018)
+------------------
 ### Added
 * Export for element responsiveTo added.