+++ packages/terra-signature/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+1.24.0 - (October 16, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+1.23.0 - (October 8, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
 1.22.0 - (September 25, 2018)
 ------------------
 ### Changed