+++ packages/terra-text/CHANGELOG.md
@@ -3,6 +3,19 @@ Changelog
 
 Unreleased
 ----------
+
+2.25.0 - (October 16, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+2.24.0 - (October 8, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+2.23.0 - (October 2, 2018)
+------------------
 ### Added
 * Exports for text font size and weight added.