+++ packages/terra-visually-hidden-text/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+1.2.0 - (October 16, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+1.1.0 - (October 8, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
 1.0.0 - (September 25, 2018)
 ------------------
 * Initial stable release